Dupla condição para aceitar deleção.

+++ src/main/java/br/aula/service/impl/CadastroServiceImpl.java
@@ -56,7 +56,7 @@ public class CadastroServiceImpl implements CadastroService{
 		CadastroEntity e = resultado.get();
 		
 		//vericar se o registro é do perfil adm. Se não for permitir deletar. Caso seja emitir mensagem de não permitido
-		if(!e.getPerfil().equals("adm")) {
+		if(!e.getPerfil().equals("adm") && !e.getAtivo()) {
 			cadastroRepository.delete(entity);	
 			foiDeletado = true;
 		}
